Add Extdata mounting support
