+++ arm9/source/filesys/fsdrive.c
@@ -8,6 +8,7 @@
 #include "image.h"
 #include "ui.h"
 #include "vff.h"
+#include "vsaveexsv.h"
 
 // last search pattern, path & mode
 static char search_pattern[256] = { 0 };
@@ -123,6 +124,8 @@ bool GetRootDirContentsWorker(DirStruct* contents) {
             snprintf(entry->name, 252, "[%s] %s (%s)", drvnum[i], drvname[i], carttype);
         else if (*(drvnum[i]) == '0') // SD card handling
             snprintf(entry->name, 252, "[%s] %s (%s)", drvnum[i], drvname[i], sdlabel);
+        else if (*(drvnum[i]) == 'F') // Save/Extdata handling
+            snprintf(entry->name, 252, "[%s] %s", drvnum[i], VSaveIsExtData() ? STR_LAB_EXTDATA_IMAGE : STR_LAB_SAVE_FILE_IMAGE);
         else snprintf(entry->name, 252, "[%s] %s", drvnum[i], drvname[i]);
         entry->size = GetTotalSpace(entry->path);
         entry->type = T_ROOT;
