+++ arm9/source/game/saveexsv.c
@@ -1,7 +1,7 @@
+#include "saveexsv.h"
 #include "common.h"
 #include "fsutil.h"
 #include "types.h"
-#include "save.h"
 #include "vff.h"
 #include "ff.h"
 
@@ -21,7 +21,7 @@ static int read_alloc_chunk(void **output, FIL *file, u64 offset, u32 size) {
 static FIL part_a = { 0 }, part_b = { 0 };
 static FIL *data_part = NULL;
 
-int SaveFileInit(SaveFile *sav) {
+int SaveExsvFileInit(SaveExsvFile *sav) {
     UINT nread = 0;
     int retval = 0;
 
@@ -45,50 +45,71 @@ int SaveFileInit(SaveFile *sav) {
     u32 dir_table_offset = 0, dir_table_size = 0;
     u32 file_table_offset = 0, file_table_size = 0;
 
-    if (fvx_read(&part_a, &sav->header, sizeof(SaveHeader), &nread) != FR_OK || nread != sizeof(SaveHeader) ||
-        sav->header.magic != 0x45564153 /* SAVE (EVAS as int) */ || sav->header.version != 0x40000 ||
-        sav->header.fs_info_offset != 0x20)
+    if (fvx_read(&part_a, &sav->pre_header, sizeof(SavePreHeader), &nread) != FR_OK || nread != sizeof(SavePreHeader))
+        goto err_exit;
+
+    u32 full_header_size = sizeof(SavePreHeader) + sizeof(SaveFsInfo);
+
+    if (sav->pre_header.magic == 0x45564153 ) { /* SAVE */
+        if (sav->pre_header.version != 0x40000 || sav->pre_header.fs_info_offset != 0x20)
+            goto err_exit;
+    } else if (sav->pre_header.magic == 0x45585356) { /* EXSV (VSXE) */
+        if (sav->pre_header.version != 0x30000 || sav->pre_header.fs_info_offset != 0x138 ||
+            fvx_read(&part_a, &sav->exsv_extra_hdr, sizeof(ExsvExtraHeader), &nread) != FR_OK || nread != sizeof(ExsvExtraHeader))
+            goto err_exit;
+
+        full_header_size += sizeof(ExsvExtraHeader);
+        sav->is_exsv = true;
+    } else {
+        goto err_exit;
+    }
+
+    if (fvx_read(&part_a, &sav->fs_info, sizeof(SaveFsInfo), &nread) != FR_OK || nread != sizeof(SaveFsInfo))
         goto err_exit;
 
     u64 part_a_size = fvx_size(&part_a);
 
     if (!sav->duplicate_meta) {
 
-        dir_table_offset = sav->header.data_region.offset + sav->header.dirtable_info.dupdata.starting_block_index * sav->header.data_region_blocksize;
-        dir_table_size = sav->header.data_region_blocksize * sav->header.dirtable_info.dupdata.block_count;
+        dir_table_offset = sav->fs_info.data_region.offset + sav->fs_info.dirtable_info.dupdata.starting_block_index * sav->fs_info.data_region_blocksize;
+        dir_table_size = sav->fs_info.data_region_blocksize * sav->fs_info.dirtable_info.dupdata.block_count;
         
-        file_table_offset = sav->header.data_region.offset + sav->header.filetable_info.dupdata.starting_block_index * sav->header.data_region_blocksize;
-        file_table_size = sav->header.data_region_blocksize * sav->header.filetable_info.dupdata.block_count;
+        file_table_offset = sav->fs_info.data_region.offset + sav->fs_info.filetable_info.dupdata.starting_block_index * sav->fs_info.data_region_blocksize;
+        file_table_size = sav->fs_info.data_region_blocksize * sav->fs_info.filetable_info.dupdata.block_count;
 
         // duplicate data has the dir and file tables in the data region, so we don't need to count them separately
-        u64 min_part_a_size = align(sizeof(SaveHeader) +
-                                    sav->header.dir_hashtbl.count * sizeof(u32) +
-                                    sav->header.file_hashtbl.count * sizeof(u32) +
-                                    (sav->header.fat.count + 1) * sizeof(SaveFatEntry),
-                                    sav->header.fs_image_blocksize) +
-                              sav->header.data_region.count * sav->header.data_region_blocksize;
+        u64 min_part_a_size = align(full_header_size +
+                                    sav->fs_info.dir_hashtbl.count * sizeof(u32) +
+                                    sav->fs_info.file_hashtbl.count * sizeof(u32) +
+                                    (sav->fs_info.fat.count + 1) * sizeof(SaveFatEntry),
+                                    sav->pre_header.fs_image_blocksize) +
+                              sav->fs_info.data_region.count * sav->fs_info.data_region_blocksize;
 
         if (part_a_size < min_part_a_size) {
-            return 1;
+            goto err_exit;
         }
     } else {
-        dir_table_offset = sav->header.dirtable_info.dupmeta.offset;
-        dir_table_size = (sav->header.dirtable_info.dupmeta.count + 2) * sizeof(SaveDirectoryEntry);
+        // it should not be possible for the extdata main DIFF to be duplicate meta
+        if (sav->is_exsv)
+            goto err_exit;
+
+        dir_table_offset = sav->fs_info.dirtable_info.dupmeta.offset;
+        dir_table_size = (sav->fs_info.dirtable_info.dupmeta.count + 2) * sizeof(SaveDirectoryEntry);
         
-        file_table_offset = sav->header.filetable_info.dupmeta.offset;
-        file_table_size = (sav->header.filetable_info.dupmeta.count + 1) * sizeof(SaveFileEntry);
+        file_table_offset = sav->fs_info.filetable_info.dupmeta.offset;
+        file_table_size = (sav->fs_info.filetable_info.dupmeta.count + 1) * sizeof(SaveFileEntry);
         
         // duplicate meta has the dir and file tables outside the data region
         u64 part_b_size = fvx_size(&part_b);
 
-        u64 min_part_a_size = align(sizeof(SaveHeader) +
-                                    sav->header.dir_hashtbl.count * sizeof(u32) +
-                                    sav->header.file_hashtbl.count * sizeof(u32) +
-                                    (sav->header.fat.count + 1) * sizeof(SaveFatEntry) +
+        u64 min_part_a_size = align(full_header_size +
+                                    sav->fs_info.dir_hashtbl.count * sizeof(u32) +
+                                    sav->fs_info.file_hashtbl.count * sizeof(u32) +
+                                    (sav->fs_info.fat.count + 1) * sizeof(SaveFatEntry) +
                                     dir_table_size +
                                     file_table_size,
-                                    sav->header.fs_image_blocksize);
-        u64 min_part_b_size = sav->header.data_region.count * sav->header.data_region_blocksize;
+                                    sav->pre_header.fs_image_blocksize);
+        u64 min_part_b_size = sav->fs_info.data_region.count * sav->fs_info.data_region_blocksize;
                               
         if (part_a_size < min_part_a_size || part_b_size < min_part_b_size)
             return 1;
@@ -100,9 +121,9 @@ int SaveFileInit(SaveFile *sav) {
     sav->max_num_file_entries = file_table_size / sizeof(SaveFileEntry);
     sav->max_num_dir_entries = dir_table_size / sizeof(SaveDirectoryEntry);
 
-    if (read_alloc_chunk((void **)&sav->dir_hashtbl, &part_a, sav->header.dir_hashtbl.offset, sav->header.dir_hashtbl.count * 4) ||
-        read_alloc_chunk((void **)&sav->file_hashtbl, &part_a, sav->header.file_hashtbl.offset, sav->header.file_hashtbl.count * 4) ||
-        read_alloc_chunk((void **)&sav->fat_entries, &part_a, sav->header.fat.offset, (sav->header.fat.count + 1) * sizeof(SaveFatEntry)) ||
+    if (read_alloc_chunk((void **)&sav->dir_hashtbl, &part_a, sav->fs_info.dir_hashtbl.offset, sav->fs_info.dir_hashtbl.count * 4) ||
+        read_alloc_chunk((void **)&sav->file_hashtbl, &part_a, sav->fs_info.file_hashtbl.offset, sav->fs_info.file_hashtbl.count * 4) ||
+        read_alloc_chunk((void **)&sav->fat_entries, &part_a, sav->fs_info.fat.offset, (sav->fs_info.fat.count + 1) * sizeof(SaveFatEntry)) ||
         read_alloc_chunk((void **)&sav->dir_entries, &part_a, dir_table_offset, dir_table_size) ||
         read_alloc_chunk((void **)&sav->file_entries, &part_a, file_table_offset, file_table_size)) {
         goto err_exit;
@@ -119,7 +140,7 @@ err_exit:
     goto ret;
 }
 
-void SaveFileFree(SaveFile *sav) {
+void SaveExsvFileFree(SaveExsvFile *sav) {
     fvx_close(&part_a);
     fvx_close(&part_b);
     data_part = NULL;
@@ -130,6 +151,7 @@ void SaveFileFree(SaveFile *sav) {
     if (sav->file_entries) { free(sav->file_entries); sav->file_entries = NULL; }
     sav->max_num_file_entries = sav->max_num_dir_entries = 0;
     sav->init_ok = false;
+    sav->is_exsv = false;
 }
 
 typedef struct ReadFileData {
@@ -167,25 +189,25 @@ static int ReadFile(u32 part_offset, u32 file_offset, u32 size, void *arg) {
 // 2: OK, early exit processing chain
 typedef int (* follow_cb)(u32 in_part_offset, u32 file_offset, u32 size, void *arg);
 
-static int ProcessFatDataBlock(SaveFile *save, u32 index, u32 *file_offset, follow_cb cb, void *cb_arg) {
+static int ProcessFatDataBlock(SaveExsvFile *save, u32 index, u32 *file_offset, follow_cb cb, void *cb_arg) {
     if (!index)
         return 0; // not yet
 
-    int cb_ret = cb(save->header.data_region.offset + (index - 1) * save->header.data_region_blocksize, *file_offset, save->header.data_region_blocksize, cb_arg);
+    int cb_ret = cb(save->fs_info.data_region.offset + (index - 1) * save->fs_info.data_region_blocksize, *file_offset, save->fs_info.data_region_blocksize, cb_arg);
     if (cb_ret)
         return cb_ret;
 
-    *file_offset += save->header.data_region_blocksize;
+    *file_offset += save->fs_info.data_region_blocksize;
     return 0;
 }
 
-static int ProcessFatChain(SaveFile *save, u32 initial_index, follow_cb cb, void *arg) {
+static int ProcessFatChain(SaveExsvFile *save, u32 initial_index, follow_cb cb, void *arg) {
     u32 idx = initial_index;
     u32 file_offset = 0;
     int ret = 0;
 
     do {
-        if (idx >= save->header.fat.count) // shouldn't happen
+        if (idx >= save->fs_info.fat.count) // shouldn't happen
             return 1;
         
         SaveFatEntry *ent = &save->fat_entries[idx];
@@ -196,7 +218,7 @@ static int ProcessFatChain(SaveFile *save, u32 initial_index, follow_cb cb, void
 
         /* process extended nodes first */
         if (ent->V.flag) {
-            if (idx + 1 >= save->header.fat.count) // same as above, shouldn't happen, but still
+            if (idx + 1 >= save->fs_info.fat.count) // same as above, shouldn't happen, but still
                 return 1;
             u32 end_index = save->fat_entries[idx + 1].V.index;
             for (u32 i = idx + 1; i < end_index + 1; i++) {
@@ -216,17 +238,21 @@ static int ProcessFatChain(SaveFile *save, u32 initial_index, follow_cb cb, void
     return 0;
 }
 
-int SaveReadFile(SaveFile *sav, u32 index, void *buffer, u32 offset, u32 count) {
+int SaveExsvReadFatFile(SaveExsvFile *sav, u32 index, void *buffer, u32 offset, u32 count) {
     if (!sav->init_ok || index >= sav->max_num_file_entries)
         return 1;
 
-    SaveFileEntry *fent = &sav->file_entries[index];
-
-    // empty file
-    if (fent->ent.first_block_index == 0x80000000)
-        return 1;
-
-    ReadFileData data = { .offset = offset, .size = count, .outbuf = (char *)buffer };
-
-    return ProcessFatChain(sav, fent->ent.first_block_index + 1, ReadFile, &data);
+    if (sav->is_exsv) {
+        return 0;
+    } else {
+        SaveFileEntry *fent = &sav->file_entries[index];
+    
+        // empty file
+        if (fent->ent.first_block_index == 0x80000000)
+            return 1;
+    
+        ReadFileData data = { .offset = offset, .size = count, .outbuf = (char *)buffer };
+    
+        return ProcessFatChain(sav, fent->ent.first_block_index + 1, ReadFile, &data);
+    }
 }
