+++ arm9/source/virtual/virtual.c
@@ -5,7 +5,7 @@
 #include "vbdri.h"
 #include "vkeydb.h"
 #include "vcart.h"
-#include "vsave.h"
+#include "vsaveexsv.h"
 #include "vvram.h"
 #include "vdisadiff.h"
 #include "ff.h"
